feat(install): restore from backup
